other: add bc_ci  (#404)

+++ tests/test_base.py
@@ -142,20 +142,30 @@ class BaseTest:
             if env_coverage.value < cls.TEST_LEVEL.value:
                 raise unittest.SkipTest(f"Skipped test : Test Coverage {env_coverage.name} < {cls.TEST_LEVEL.name}")
 
-            if os.path.exists(cls.get_rbln_local_dir()):
-                shutil.rmtree(cls.get_rbln_local_dir())
+            REUSE_ARTIFACTS_PATH = os.environ.get("REUSE_ARTIFACTS_PATH", None)
+            if REUSE_ARTIFACTS_PATH is None:
+                if os.path.exists(cls.get_rbln_local_dir()):
+                    shutil.rmtree(cls.get_rbln_local_dir())
 
-            cls.model = cls.RBLN_CLASS.from_pretrained(
-                cls.HF_MODEL_ID,
-                model_save_dir=cls.get_rbln_local_dir(),
-                rbln_device=cls.DEVICE,
-                **cls.RBLN_CLASS_KWARGS,
-                **cls.HF_CONFIG_KWARGS,
-            )
+                cls.model = cls.RBLN_CLASS.from_pretrained(
+                    cls.HF_MODEL_ID,
+                    model_save_dir=cls.get_rbln_local_dir(),
+                    rbln_device=cls.DEVICE,
+                    **cls.RBLN_CLASS_KWARGS,
+                    **cls.HF_CONFIG_KWARGS,
+                )
+            else:
+                if os.path.exists(REUSE_ARTIFACTS_PATH):
+                    compiled_model_path = os.path.join(REUSE_ARTIFACTS_PATH, cls.get_rbln_local_dir())
+                    if os.path.exists(compiled_model_path):
+                        with ContextRblnConfig(device=-1):
+                            cls.model = cls.RBLN_CLASS.from_pretrained(compiled_model_path)
+                if not hasattr(cls, "model"):
+                    raise unittest.SkipTest("Compiled model not found")
 
         @classmethod
         def get_rbln_local_dir(cls):
-            return os.path.basename(cls.HF_MODEL_ID) + "-local"
+            return os.path.basename(cls.__module__.split(".")[-1]) + "_" + os.path.basename(cls.__name__) + "-artifact"
 
         @classmethod
         def get_hf_auto_class(cls):
@@ -175,6 +185,22 @@ class BaseTest:
         def tearDownClass(cls):
             if os.path.exists(cls.get_rbln_local_dir()):
                 shutil.rmtree(cls.get_rbln_local_dir())
+
+        # BC: Test save_artifacts and copy tree is successful
+        def test_save_artifacts(self):
+            SAVE_ARTIFACTS_PATH = os.environ.get("SAVE_ARTIFACTS_PATH", None)
+            if SAVE_ARTIFACTS_PATH is None:
+                return
+            else:
+                os.makedirs(SAVE_ARTIFACTS_PATH, exist_ok=True)
+                saved_path = os.path.join(SAVE_ARTIFACTS_PATH, self.get_rbln_local_dir())
+                shutil.copytree(self.get_rbln_local_dir(), saved_path, dirs_exist_ok=True)
+
+                with ContextRblnConfig(create_runtimes=False):
+                    _ = self.RBLN_CLASS.from_pretrained(
+                        saved_path,
+                        **self.HF_CONFIG_KWARGS,
+                    )
 
         def test_model_save_dir(self):
             self.assertTrue(os.path.exists(self.get_rbln_local_dir()), "model_save_dir does not work.")
@@ -197,7 +223,8 @@ class BaseTest:
                     output = self.model(**inputs)[0]
 
             output = self.postprocess(inputs, output)
-            if self.EXPECTED_OUTPUT and self.DEVICE is None:
+            REUSE_ARTIFACTS_PATH = os.environ.get("REUSE_ARTIFACTS_PATH", None)
+            if self.EXPECTED_OUTPUT and self.DEVICE is None and REUSE_ARTIFACTS_PATH is None:
                 from simphile import jaccard_similarity
 
                 if isinstance(self.EXPECTED_OUTPUT, str):
@@ -240,10 +267,11 @@ class BaseTest:
                 self._inner_test_save_load(tmpdir)
 
         def test_model_save_dir_load(self):
+            rbln_local_dir = self.get_rbln_local_dir()
             with ContextRblnConfig(create_runtimes=False):
                 # Test model_save_dir
                 _ = self.RBLN_CLASS.from_pretrained(
-                    self.get_rbln_local_dir(),
+                    rbln_local_dir,
                     rbln_create_runtimes=False,
                     **self.HF_CONFIG_KWARGS,
                 )
